// src/modules/gravity-forms/generated/graphql.ts
export type Maybe<T> = T | null;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
export type MakeOptional<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]?: Maybe<T[SubKey]> };
export type MakeMaybe<T, K extends keyof T> = Omit<T, K> & { [SubKey in K]: Maybe<T[SubKey]> };
/** All built-in and custom scalars, mapped to their actual values */
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
};
// ...
export type ChoiceProperty = {
  __typename?: 'ChoiceProperty';
  /** Determines if this choice should be selected by default when displayed. The value true will select the choice, whereas false will display it unselected. */
  isSelected?: Maybe<Scalars['Boolean']>;
  /** The text to be displayed to the user when displaying this choice. */
  text?: Maybe<Scalars['String']>;
  /** The value to be stored in the database when this choice is selected. Note: This property is only supported by the Drop Down and Post Category fields. Checkboxes and Radio fields will store the text property in the database regardless of the value property. */
  value?: Maybe<Scalars['String']>;
};
// ...
export type ConditionalLogic = {
  __typename?: 'ConditionalLogic';
  /** The type of action the conditional logic will perform. */
  actionType?: Maybe<ConditionalLogicActionTypeEnum>;
  /** Determines how to the rules should be evaluated. */
  logicType?: Maybe<ConditionalLogicLogicTypeEnum>;
  /** Conditional logic rules. */
  rules?: Maybe<Array<Maybe<ConditionalLogicRule>>>;
};

/** The type of action the conditional logic will perform. */
export enum ConditionalLogicActionTypeEnum {
  /** Text button (default). */
  Hide = 'HIDE',
  /** Image button. */
  Show = 'SHOW',
}

/** Determines how to the rules should be evaluated. */
export enum ConditionalLogicLogicTypeEnum {
  /** Evaulate all logic rules. */
  All = 'ALL',
  /** Evaluate any logic rule. */
  Any = 'ANY',
}

/** Gravity Forms conditional logic rule. */
export type ConditionalLogicRule = {
  __typename?: 'ConditionalLogicRule';
  /** Target field Id. Field that will have it’s value compared with the value property to determine if this rule is a match. */
  fieldId?: Maybe<Scalars['Float']>;
  /** Operator to be used when evaluating this rule. */
  operator?: Maybe<RuleOperatorEnum>;
  /** The value to compare with field specified by fieldId. */
  value?: Maybe<Scalars['String']>;
};
// ...
export enum DescriptionPlacementPropertyEnum {
  /** The field description is displayed above the field input (i.e. immediately after the field label) */
  Above = 'ABOVE',
  /** The field description is displayed below the field input. */
  Below = 'BELOW',
  /** The field description is inherited from the form default settings. */
  Inherit = 'INHERIT',
}
// ...
export type FieldError = {
  __typename?: 'FieldError';
  /** The field with the associated error message */
  id?: Maybe<Scalars['Float']>;
  /** Error message. */
  message?: Maybe<Scalars['String']>;
};
// ...
export type FormField = {
  /** Controls the visibility of the field based on values selected by the user. */
  conditionalLogic?: Maybe<ConditionalLogic>;
  /** String containing the custom CSS classes to be added to the <li> tag that contains the field. Useful for applying custom formatting to specific fields. */
  cssClass?: Maybe<Scalars['String']>;
  /**
   * Array of the custom CSS classes to be added to the <li> tag that contains the field. Useful for applying custom formatting to specific fields.
   * @deprecated Please use `cssClass` instead.
   */
  cssClassList?: Maybe<Array<Maybe<Scalars['String']>>>;
  /** The ID of the form this field belongs to. */
  formId: Scalars['Int'];
  /** Field ID. */
  id: Scalars['Int'];
  databaseId: Scalars['Int'];
  /** The type of field to be displayed. */
  type: Scalars['String'];
};
// ...
export type MultiSelectField = FormField & {
  __typename?: 'MultiSelectField';
  /** When specified, the value of this property will be used on the admin pages instead of the label. It is useful for fields with long labels. */
  adminLabel?: Maybe<Scalars['String']>;
  /** Determines if this field should only visible on the administration pages. A value of 1 will mark the field as admin only and will hide it from the public form. Useful for fields such as “status” that help with managing entries, but don’t apply to users filling out the form. */
  adminOnly?: Maybe<Scalars['Boolean']>;
  /** Determines if the field’s value can be pre-populated dynamically. 1 to allow field to be pre-populated, 0 otherwise. */
  allowsPrepopulate?: Maybe<Scalars['Boolean']>;
  /** Contains the available choices for the field. For instance, drop down items and checkbox items are configured with this property. */
  choices?: Maybe<Array<Maybe<ChoiceProperty>>>;
  /** Controls the visibility of the field based on values selected by the user. */
  conditionalLogic?: Maybe<ConditionalLogic>;
  /** String containing the custom CSS classes to be added to the &lt;li&gt; tag that contains the field. Useful for applying custom formatting to specific fields. */
  cssClass?: Maybe<Scalars['String']>;
  /**
   * Array of the custom CSS classes to be added to the &lt;li&gt; tag that contains the field. Useful for applying custom formatting to specific fields.
   * @deprecated Please use `cssClass` instead.
   */
  cssClassList?: Maybe<Array<Maybe<Scalars['String']>>>;
  /** Field description. */
  description?: Maybe<Scalars['String']>;
  /** The placement of the field description. */
  descriptionPlacement?: Maybe<DescriptionPlacementPropertyEnum>;
  /** Determines if the field (checkbox, select or radio) have choice values enabled, which allows the field to have choice values different from the labels that are displayed to the user. */
  enableChoiceValue?: Maybe<Scalars['Boolean']>;
  /** When set to true, the &quot;Chosen&quot; jQuery script will be applied to this field, enabling search capabilities to Drop Down fields and a more user-friendly interface for Multi Select fields. */
  enableEnhancedUI?: Maybe<Scalars['Boolean']>;
  /** Contains the message that is displayed for fields that fail validation. */
  errorMessage?: Maybe<Scalars['String']>;
  /** The ID of the form this field belongs to. */
  formId: Scalars['Int'];
  /** Field ID. */
  id: Scalars['Int'];
  /** Assigns a name to this field so that it can be populated dynamically via this input name. Only applicable when allowsPrepopulate is set to 1. */
  inputName?: Maybe<Scalars['String']>;
  /** Determines if the field requires the user to enter a value. 1 marks the field as required, 0 marks the field as not required. Fields marked as required will prevent the form from being submitted if the user has not entered a value in it. */
  isRequired?: Maybe<Scalars['Boolean']>;
  /** Field label that will be displayed on the form and on the admin pages. */
  label?: Maybe<Scalars['String']>;
  /** Determines the size of the field when displayed on the page. */
  size?: Maybe<SizePropertyEnum>;
  /** The type of field to be displayed. */
  type: Scalars['String'];
  /** MultiSelect field value. */
  values?: Maybe<Array<Maybe<Scalars['String']>>>;
  /** Field visibility. */
  visibility?: Maybe<VisibilityPropertyEnum>;
};
// ...
export enum SizePropertyEnum {
  /** Large field size. */
  Large = 'LARGE',
  /** Medium field size. */
  Medium = 'MEDIUM',
  /** Small field size. */
  Small = 'SMALL',
}
// ...
export enum VisibilityPropertyEnum {
  /** The field is for "administrative" use. */
  Administrative = 'ADMINISTRATIVE',
  /** The field is "hidden". */
  Hidden = 'HIDDEN',
  /** The field is "visible". */
  Visible = 'VISIBLE',
}